refactor: Context 파일 Provider로 변경
- modalContext, snackbarContext 파일명을 Provider로 변경
- 툴팁 컴포넌트 내 스타일 관련 로직 개선 (src/headless-ui/tooltip/index.tsx)
