+++ src/headless-ui/_stories/snackbar.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react'
 import { useRef } from 'react'
 
-import { SnackbarProvider, useSetSnackbar } from '@/headless-ui/snackbar/snackbarContext'
+import { SnackbarProvider, useSetSnackbar } from '@/headless-ui/snackbar/snackbarProvider'
 
 const SnackbarItem = ({ 
   onClose,
